requirements.txt
fastapi
uvicorn
openai
langchain
python-dotenv
requests
pymongo
sentence-transformers
chromadb
tqdm

